ai/pyproject.toml
[project]
name = "app"
version = "1.0.0"
description = ""
readme = "README.md"
requires-python = ">=3.11, <3.12"

dependencies = [
    "fastapi[all] (>=0.115.12,<0.116.0)",
    "numpy (>=1.24,<2.0)",
    "facenet-pytorch (>=2.6.0,<3.0.0)",
    "torchvision (==0.17.0)",
    "scipy (>=1.15.3,<2.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "ultralytics (>=8.3.130,<9.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "redis (>=6.0.0,<7.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    "pillow-heif (>=0.22.0,<0.23.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "python-socketio[asgi] (>=5.13.0,<6.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "livekit-api (>=1.0.2,<2.0.0)",
    "imagehash (>=4.3.2,<5.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.12"
black = "^25.1.0"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

